+++ src/main/java/org/yandrut/pages/CalculatorPage.java
@@ -7,16 +7,62 @@ import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.yandrut.utils.TestListener;
 
+import java.util.List;
+
 public class CalculatorPage extends AbstractPage {
+
+    @FindBy(xpath = "//span[text()='Add to estimate']")
+    private WebElement addToEstimate;
+    @FindBy(xpath = "//h2[text()='Compute Engine']/..")
+    private WebElement computeEngine;
+
+    @FindBy(xpath = "//div[@data-field-input-type='2' and @data-field-type='158']" )
+    private WebElement gpuModelDropdown;
+    @FindBy(xpath = "//ul[@aria-label='Region']/../../../../../..")
+    private WebElement locationDropdown;
+    @FindBy(xpath = "//ul[@aria-label='Local SSD']/../../../../..")
+    private WebElement localSsdDropdown;
+    @FindBy(xpath = "//ul[@aria-label='Machine type']/../../..")
+    private WebElement machineTypeDropdown;
+
+    @FindBy(xpath = "//ul[@aria-label='GPU Model']")
+    private List<WebElement> gpuModelList;
+    @FindBy(xpath = "//ul[@aria-label='Local SSD']/li")
+    private List<WebElement> localSsdList;
+    @FindBy(xpath = "//ul[@aria-label='Region']/li")
+    private List<WebElement> locationsList;
+    @FindBy(xpath = "//ul[@aria-label='Machine type']/li")
+    private List<WebElement> machineTypesList;
+
+    @FindBy(xpath = "//input[@max='50000']")
+    private WebElement numberOfInstancesInput;
+    @FindBy(xpath = "//button[@aria-label='Add GPUs']")
+    private WebElement addGpu;
+    @FindBy(xpath = "//label[text()='1 year']/..")
+    private WebElement commitedUsage;
+    @FindBy(xpath = "//div[text()='Estimated cost']/following::label")
+    private WebElement estimatedCost;
+
+    private final String suffixForTextLocators = "/../..//div/span/following-sibling::span/span";
+    private final String machineTextSelectorString = "//ul[@aria-label='Machine type']" + suffixForTextLocators;
+    @FindBy(xpath = machineTextSelectorString)
+    private WebElement machineTypeText;
+
+    private final String gpuTextSelectorString = "//ul[@aria-label='GPU Model']" + suffixForTextLocators;
+    @FindBy(xpath = gpuTextSelectorString)
+    private WebElement gpuModelText;
+
+    private final String localSSDTextLocatorString = "//ul[@aria-label='Local SSD']" + suffixForTextLocators;
+    @FindBy(xpath = localSSDTextLocatorString)
+    private WebElement localSSDText;
+
+    private final String regionTextLocatorString = "//ul[@aria-label='Region']" + suffixForTextLocators;
+    @FindBy(xpath = regionTextLocatorString)
+    private WebElement regionTextLocator;
+
     private final WebDriver driver;
     private static final Logger log = LogManager.getLogger(CalculatorPage.class);
 
-    @FindBy(xpath = "//button[@data-idom-class='xhASFc']")
-    private WebElement addToEstimate;
-    @FindBy(xpath = "//div[@data-service-form='8']")
-    private WebElement computeEngine;
-    @FindBy(xpath = "//div[text()='Estimated cost']/following::label")
-    private WebElement estimatedCost;
 
     public CalculatorPage(WebDriver driver) {
         super(driver);
@@ -30,10 +76,6 @@ public class CalculatorPage extends AbstractPage {
         return this;
     }
 
-    public CalculatorFormLoader fillOutTheForm() {
-        return new CalculatorFormLoader(driver);
-    }
-
     public CalculatorPage clickOnComputeEngine() {
         addToEstimate.click();
         computeEngine.click();
@@ -41,10 +83,58 @@ public class CalculatorPage extends AbstractPage {
         return this;
     }
 
+    public CalculatorPage setNumberOfInstances(String numberOfInstances) {
+        numberOfInstancesInput.clear();
+        numberOfInstancesInput.sendKeys(numberOfInstances);
+        log.info("Actual number of instances is: {}", numberOfInstances);
+        return this;
+    }
+
+    public CalculatorPage setMachineType(String machineType) {
+        machineTypeDropdown.click();
+        selectFromDropdownList(machineTypesList, machineType);
+        log.info("Actual machine type is: {} \n Expected type: {}", machineTypeText.getText(), machineType);
+        return this;
+    }
+
+    public CalculatorPage addGpuModel(String gpuModel) {
+        addGpu.click();
+        gpuModelDropdown.click();
+        selectFromDropdownList(gpuModelList, gpuModel);
+        log.info("Actual GPU model is: {} \n Expected model: {}", gpuModelText.getText(), gpuModel);
+        return this;
+    }
+
+    public CalculatorPage setLocalSSD(String ssdName) {
+        localSsdDropdown.click();
+        selectFromDropdownList(localSsdList, ssdName);
+        log.info("Actual local SSD is: {} \n Expected is: {}", localSSDText.getText(), ssdName);
+        return this;
+    }
+
+    public CalculatorPage setLocation(String location) {
+        locationDropdown.click();
+        selectFromDropdownList(locationsList, location);
+        log.info("Actual location: {} \n Expected is: {}", regionTextLocator.getText(), location);
+        return this;
+    }
+
+    public void clickOnCommitedUsage() {
+        commitedUsage.click();
+    }
+
     public String getEstimatedCost() {
         String priceText = estimatedCost.getText();
         log.info("Actual price is {}", priceText);
-        new TestListener().saveScreenshot();
+        TestListener.saveScreenshots();
         return priceText;
+    }
+
+    private void selectFromDropdownList(List<WebElement> dropdownList, String dropdownElementName) {
+        dropdownList.stream()
+                .filter((e) -> e.getText().equals(dropdownElementName))
+                .findAny()
+                .ifPresent(WebElement::click);
+        log.info("Choosing an option: {}", dropdownElementName);
     }
 }
