+++ src/main/resources/main.properties
@@ -1,6 +1,9 @@
 data.form.NUMBER_OF_INSTANCES = 4
 data.form.MACHINE_TYPE = n1-standart-8
-data.form.GPU_TYPE = NVIDIA Tesla V100
-data.form.LOCAL_SSD = 2x375 Gb;
-data.form.DATA_CENTER_LOCATION = Frankfurt (europe-west3);
-data.form.ESTIMATED_COST = 0
+data.form.GPU_TYPE = NVIDIA V100
+data.form.LOCAL_SSD = 2x375 GB
+data.form.DATA_CENTER_LOCATION = Netherlands (europe-west4)
+data.form.ESTIMATED_COST = $1,820.80
+data.provided.cloudUrl = https://cloud.google.com
+data.provided.calculatorUrl = https://cloud.google.com/products/calculator
+data.provided.searchPrompt = Google Cloud Pricing Calculator
